Create separate system::Null type.
