+++ src/internal/system/mod.rs
@@ -1,1 +1,5 @@
 pub(crate) mod schedule;
+
+mod null;
+
+pub(crate) use null::Null;
